also fixed some code issues and fully RE'd PlayerObject::resetObject

+++ Classes/ColoredSection.h
@@ -9,7 +9,7 @@ public:
 	bool init(cocos2d::ccColor3B col, int startPos, int endPos);
 
 	CC_SYNTHESIZE(cocos2d::ccColor3B, m_col, Col); // 0x18
-	CC_SYNTHESIZE(float, m_startPos, StartPos); // 0x1c
+	CC_SYNTHESIZE(int, m_startPos, StartPos); // 0x1c
 	CC_SYNTHESIZE(int, m_endPos, EndPos); // 0x20
 
 };
